refactor(dashboard): shared row fixture and sorting polish
