+++ src/renderer/features/dashboard-accounts-table/lib/sorting.ts
@@ -9,12 +9,15 @@ export const nextSort = (current: TableSortState, key: SortKey): TableSortState 
 };
 
 export const sortRows = (rows: AccountRow[], sort: TableSortState): AccountRow[] => {
-  const mul = sort.dir === 'asc' ? 1 : -1;
+  const multiplier = sort.dir === 'asc' ? 1 : -1;
 
   return [...rows].sort((a, b) => {
-    if (sort.key === 'chain') return mul * a.chain.name.localeCompare(b.chain.name);
+    if (sort.key === 'chain') return multiplier * a.chain.name.localeCompare(b.chain.name);
 
-    return mul * ((a.fiat[sort.key] ?? Number.NEGATIVE_INFINITY) - (b.fiat[sort.key] ?? Number.NEGATIVE_INFINITY));
+    // null fiat sorts as the lowest value (not pinned to the end), so it flips with direction
+    return (
+      multiplier * ((a.fiat[sort.key] ?? Number.NEGATIVE_INFINITY) - (b.fiat[sort.key] ?? Number.NEGATIVE_INFINITY))
+    );
   });
 };
 
@@ -23,6 +26,7 @@ export type GroupOrder = 'value' | 'name';
 export const sortGroups = (groups: AccountGroup[], order: GroupOrder): AccountGroup[] => {
   if (order === 'name') return [...groups].sort((a, b) => a.name.localeCompare(b.name));
 
+  // same null-as-lowest-value rule as sortRows above
   return [...groups].sort(
     (a, b) => (b.subtotalFiat ?? Number.NEGATIVE_INFINITY) - (a.subtotalFiat ?? Number.NEGATIVE_INFINITY),
   );
